requirements.txt
#CORS
django-cors-headers

#text editor
django-ckeditor
pillow